bugfix in border cropping
Previously, when cropping borders to the context size, we checked that
the border coordinates were right up against the context boundaries.
This isn't right, because widgets are free to use less than the amount
of space available in the context. This commit relaxes the checks to
merely guarantee that the border coordinates are in bounds.